databasefeeder/requirements.txt
kafka-python==1.4.4
asgiref==2.3.2
PyJWT==1.7.1
pymongo==3.4.0
avro-python3==1.8.2
requests==2.21.0